double -> float32

+++ phils_model/src/activations.c
@@ -8,7 +8,7 @@
 // Activation functions
 ///////////////////////////////////////////////////////////////////////////////
 
-void relu_calc(double **y, int matrix_rows, int matrix_columns) {
+void relu_calc(float **y, int matrix_rows, int matrix_columns) {
     for (int i = 0; i < matrix_rows; ++i) {
         for (int j = 0; j < matrix_columns; ++j) {
             if (y[i] > 0) {
@@ -21,7 +21,7 @@ void relu_calc(double **y, int matrix_rows, int matrix_columns) {
     }
 }
 
-void relu_derivative(double **y, int matrix_rows, int matrix_columns) {
+void relu_derivative(float **y, int matrix_rows, int matrix_columns) {
     for (int i = 0; i < matrix_rows; ++i) {
         for (int j = 0; j < matrix_columns; ++j) {
             if (y[i] > 0) {
@@ -36,8 +36,8 @@ void relu_derivative(double **y, int matrix_rows, int matrix_columns) {
 
 ///////////////////////////////////////////////////////////////////////////////
 
-double sigmoid(double x) {
-    double n = exp(x);
+float sigmoid(float x) {
+    float n = exp(x);
     if (x >= 0) {
         return 1.0 / (1.0 + n);
     } else {
@@ -45,7 +45,7 @@ double sigmoid(double x) {
     }
 }
 
-void sigmoid_calc(double **y, int matrix_rows, int matrix_columns) {
+void sigmoid_calc(float **y, int matrix_rows, int matrix_columns) {
     for (int i = 0; i < matrix_rows; ++i) {
         for (int j = 0; j < matrix_columns; ++j) {
             y[i][j] = sigmoid(y[i][j]);
@@ -53,8 +53,8 @@ void sigmoid_calc(double **y, int matrix_rows, int matrix_columns) {
     }
 }
 
-void sigmoid_derivative(double **y, int matrix_rows, int matrix_columns) {
-    double **f = create_matrix(matrix_rows, matrix_columns);
+void sigmoid_derivative(float **y, int matrix_rows, int matrix_columns) {
+    float **f = create_matrix(matrix_rows, matrix_columns);
 
     for (int i = 0; i < matrix_rows; ++i) {
         for (int j = 0; j < matrix_columns; ++j) {
@@ -72,8 +72,8 @@ void sigmoid_derivative(double **y, int matrix_rows, int matrix_columns) {
 ///////////////////////////////////////////////////////////////////////////////
 
 // The softmax method (returns normalized class probabilities)
-void softmax_calc(double **y, int matrix_rows, int matrix_columns) {
-    double max_val = y[0][0];
+void softmax_calc(float **y, int matrix_rows, int matrix_columns) {
+    float max_val = y[0][0];
 
     for (int i = 1; i < matrix_rows; ++i) {
         for (int j = 1; j < matrix_columns; ++j) {
@@ -84,7 +84,7 @@ void softmax_calc(double **y, int matrix_rows, int matrix_columns) {
     }
 
     // Let's subtract the maximum from each element to stabilize the exponent
-    double sum_exp = 0.0;
+    float sum_exp = 0.0;
     for (int i = 0; i < matrix_rows; ++i) {
         for (int j = 0; j < matrix_columns; ++j) {
             y[i][j] = exp(y[i][j] - max_val);
@@ -100,7 +100,7 @@ void softmax_calc(double **y, int matrix_rows, int matrix_columns) {
     }
 }
 
-void softmax_derivative(double **y, int matrix_rows, int matrix_columns) {
+void softmax_derivative(float **y, int matrix_rows, int matrix_columns) {
     for (int i = 0; i < matrix_rows; ++i) {
         for (int j = 0; j < matrix_columns; ++j) {
             y[i][j] = y[i][j];
@@ -110,7 +110,7 @@ void softmax_derivative(double **y, int matrix_rows, int matrix_columns) {
 
 ///////////////////////////////////////////////////////////////////////////////
 
-void apply_activation_calc(double **y, int matrix_rows, int matrix_columns, int activation) {
+void apply_activation_calc(float **y, int matrix_rows, int matrix_columns, int activation) {
     if (activation == 0) {
         relu_calc(y, matrix_rows, matrix_columns);
     } else if (activation == 1) {
@@ -122,7 +122,7 @@ void apply_activation_calc(double **y, int matrix_rows, int matrix_columns, int 
     }
 }
 
-void apply_activation_derivative(double **y, int matrix_rows, int matrix_columns, int activation) {
+void apply_activation_derivative(float **y, int matrix_rows, int matrix_columns, int activation) {
     if (activation == 0) {
         relu_derivative(y, matrix_rows, matrix_columns);
     } else if (activation == 1) {
