+++ phils_model/src/adam.h
@@ -3,38 +3,38 @@
 
 
 typedef struct AdamOptimizer {
-    double ***m;    // Накопленный средний градиент
-    double ***v;    // Накопленный квадратный градиент
-    double lr;      // Скорость обучения
-    double b1;      // Коэффициент для среднего градиента
-    double b2;      // Коэффициент для квадрата градиента
-    double eps;     // Маленькое значение для избежания деления на ноль
+    float ***m;    // Накопленный средний градиент
+    float ***v;    // Накопленный квадратный градиент
+    float lr;      // Скорость обучения
+    float b1;      // Коэффициент для среднего градиента
+    float b2;      // Коэффициент для квадрата градиента
+    float eps;     // Маленькое значение для избежания деления на ноль
     int epoch;          // Текущая эпоха
 } AdamOptimize;
 
 
 void destroy_adam(
     struct AdamOptimizer *opt,
-    double *layer_sizes,
+    float *layer_sizes,
     int layer_sizes_rows,
     int layer_sizes_cols
 );
 
 struct AdamOptimizer *create_adam(
-    double lr,
-    double b1,
-    double b2,
-    double eps,
-    double *layer_sizes,
+    float lr,
+    float b1,
+    float b2,
+    float eps,
+    float *layer_sizes,
     int layer_sizes_rows,
     int layer_sizes_cols
 );
 
 void adam_step(
     struct AdamOptimizer *optimizer,
-    double ***weights,
-    double ***grads,
-    double *layer_sizes,
+    float ***weights,
+    float ***grads,
+    float *layer_sizes,
     int layer_sizes_rows,
     int layer_sizes_cols
 );
