+++ phils_model/src/backward.h
@@ -3,16 +3,16 @@
 
 
 typedef struct BackwardData {
-    double ***weights;
-    double ****X_list;
-    double ****Y_list;
-    double ****grad_w_list;
-    double ****grad_x_list;
-    double ***grad_b_list;
-    double **targets;
-    double *layer_sizes;
-    double *activations;
-    double *epoch_losses;
+    float ***weights;
+    float ****X_list;
+    float ****Y_list;
+    float ****grad_w_list;
+    float ****grad_x_list;
+    float ***grad_b_list;
+    float **targets;
+    float *layer_sizes;
+    float *activations;
+    float *epoch_losses;
     int start_idx;
     int end_idx;
     int layer_sizes_rows;
@@ -31,24 +31,24 @@ void *backward_worker(
 
 void backward_threading(
     struct BackwardData backward_thread_data[],
-    double ***weights,
-    double **targets,
-    double **biases,
-    double ****X_list,
-    double ****Y_list,
-    double ****grad_w_list,
-    double ****grad_x_list,
-    double ***grad_b_list,
-    double *layer_sizes,
+    float ***weights,
+    float **targets,
+    float **biases,
+    float ****X_list,
+    float ****Y_list,
+    float ****grad_w_list,
+    float ****grad_x_list,
+    float ***grad_b_list,
+    float *layer_sizes,
     int layer_sizes_rows,
     int layer_sizes_cols,
     int dataset_samples_rows,
     int dataset_samples_cols,
     int dataset_targets_cols,
     int matrix_rows,
-    double *activations,
+    float *activations,
     int loss,
-    double *epoch_losses,
+    float *epoch_losses,
     int regression,
     int num_threads
 );
